backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
pandas
numpy
shap
openai
httpx
python-dotenv
pytest
pytest-asyncio
